Finish parity follow-ups: timed NPC specials, grazing, jail, tithing, item damage
Region RECALLIN/MARK read and toggle REGION_ANTIMAGIC_RECALL_IN with the
Source-X polarity (recall into any region was refused); the Teleport
spell target point survives CastDone.

NPC breath and throw run a start stage and resolve after 3 s with the
LOS re-check / current target position; grazing uses t_grass typedef
terrain plus the region grass resource; OnHarmedBy reacts before armor
with the 1-in-10 switch; NPC attacks run CheckCrimeSeen; flee turns 3-5
random dirs; NPC spell lists merge all spellbooks.

Skills: ground-item steal on a failed roll and witnessed-theft karma,
karma change messages, lockpick fail @Damage, provocation on real noto
and crime checks, smelt gem/consume branches, campfire LIGHT_LARGE, tool
wear after @Success with ITEMDAMAGECHANCE/AMOUNT locals.

Magic: tithing cost via @Select/TithingLoss, FollowerSlotsOverride,
@Fail fizzle effect locals, MOVE_NEVER effects survive low-level dispel
and death, jailed recall/teleport redirect and gate refusal.

Items: shared CItem::OnTakeDamage port (armor/weapon wear, SELFREPAIR,
arrows, webs, explosion potions, @Damage veto), 64-bit ATTR with
ATTR_CANNOTREPAIR, @Drink RETURN 5/6, ENHANCEPOTIONS.

Characters: @Jailed before jail/forgive with cancel, JAIL verb cell,
InitPlayer hue and hair/beard validation per race.

+++ src/SphereNet.Game/Objects/Items/Item.cs
@@ -2220,9 +2220,9 @@ public class Item : ObjBase
             }
             // IC_NODROP / IC_NOTRADE / IC_QUESTITEM (CItem.cpp:2786-2794): the masked
             // attribute bit itself, as a decimal number (0 when clear).
-            case "NODROP": value = ((uint)Attributes & AttrNoDrop).ToString(); return true;
-            case "NOTRADE": value = ((uint)Attributes & AttrNoTrade).ToString(); return true;
-            case "QUESTITEM": value = ((uint)Attributes & AttrQuestItem).ToString(); return true;
+            case "NODROP": value = ((ulong)Attributes & AttrNoDrop).ToString(); return true;
+            case "NOTRADE": value = ((ulong)Attributes & AttrNoTrade).ToString(); return true;
+            case "QUESTITEM": value = ((ulong)Attributes & AttrQuestItem).ToString(); return true;
             // IC_TOPCONT (CItem.cpp:2804 -> GetTopContainer, :4162): the outermost
             // CONTAINER ITEM holding this one, stopping below a character; 0 when the
             // item is not inside a container at all.
@@ -3369,8 +3369,8 @@ public class Item : ObjBase
                 };
                 bool on = string.IsNullOrWhiteSpace(value) || ParseBaseDefNumber(value) != 0;
                 Attributes = on
-                    ? (ObjAttributes)((uint)Attributes | flag)
-                    : (ObjAttributes)((uint)Attributes & ~flag);
+                    ? (ObjAttributes)((ulong)Attributes | flag)
+                    : (ObjAttributes)((ulong)Attributes & ~(ulong)flag);
                 return true;
             }
             // The door keys are SetDefNum'd numbers (CItem.cpp:3167-3170). DOOROPENID
